Moving templated functions to their own files
